DependencyNavigator: Support querying sub projects
Move the functionality to query the sub projects of a project from
Project to DependencyNavigator. The implementation in Project is based
on the scopes structure and thus has to be replaced. The new
implementation can be based on the already existing
scopeDependencies() functions.

While at it, store the dependency navigator in OrtResult in a lazy
property, to prevent the creation of a new instance at each access.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>