DependencyNavigator: Support calculating the dependency tree depth
This functionality is required by StatisticsCalculator. Adapt the
calculator to use the new functions of DependencyNavigator.

The functionality can already be implemented in a generic way in the
DependencyNavigator interface. Remove the no longer needed
getDependencyTreeDepth() function from Scope. As this is the only
function that was tested in ScopeTest, the test class can be removed,
too.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>