Rework usages of Project.collectDependencies()
When using the dependency graph format a project's dependencies are no
longer stored in the project itself. Therefore, code that needs this
information has to be changed to query the result's
DependencyNavigator instead.

This commit deals with client code which is uncritical. There are a
few places left that need to be addressed in a special way.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>