DependencyNavigator: Extract an abstract test base class
The base class defines the test cases for the single functions of the
DependencyNavigator interface. Concrete sub classes have to provide
result files with the expected data in the correct format, so that the
implementation to be tested is used. That way the same test setup can
be used for different implementations, and the results can be compared
directly.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>